Add files via upload
Fixed majority of issues with implementing header and cpp files. Running into errors in line 32 and 33 of main that drivetrain.ISR_countL() and drivetrain.ISR_countR(): argument of type "void" is incompatible with parameter of type "void (*)()"

The compiler also cannot identify the abs() function in mapping regardless of if I import the cmath library.